Refactoring cobalt

+++ modules/lib/src/main/java/com/github/auties00/cobalt/wam/event/AboutConsumptionDailyEvent.java
@@ -8,20 +8,20 @@ import com.github.auties00.cobalt.wam.annotation.WamProperty;
 import com.github.auties00.cobalt.wam.model.WamType;
 
 import java.util.Optional;
-import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 @WhatsAppWebModule(moduleName = "WAWebAboutConsumptionDailyWamEvent")
 @WamEvent(id = 6816)
 public interface AboutConsumptionDailyEvent extends WamEventSpec {
     @WamProperty(index = 1, type = WamType.INTEGER)
-    OptionalInt aboutChatBubbleTapCount();
+    OptionalLong aboutChatBubbleTapCount();
 
     @WamProperty(index = 2, type = WamType.INTEGER)
-    OptionalInt aboutChatConsumptionCount();
+    OptionalLong aboutChatConsumptionCount();
 
     @WamProperty(index = 3, type = WamType.STRING)
     Optional<String> aboutLocale();
 
     @WamProperty(index = 4, type = WamType.INTEGER)
-    OptionalInt aboutMessageSendCount();
+    OptionalLong aboutMessageSendCount();
 }
