+++ modules/lib/src/main/java/com/github/auties00/cobalt/wam/event/AboutCreationEvent.java
@@ -13,19 +13,19 @@ import com.github.auties00.cobalt.wam.type.PresetType;
 
 import java.time.Instant;
 import java.util.Optional;
-import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 @WhatsAppWebModule(moduleName = "WAWebAboutCreationWamEvent")
 @WamEvent(id = 6818)
 public interface AboutCreationEvent extends WamEventSpec {
     @WamProperty(index = 1, type = WamType.INTEGER)
-    OptionalInt aboutDuration();
+    OptionalLong aboutDuration();
 
     @WamProperty(index = 2, type = WamType.ENUM)
     Optional<AboutEntrypointType> aboutEntrypoint();
 
     @WamProperty(index = 3, type = WamType.INTEGER)
-    OptionalInt aboutLength();
+    OptionalLong aboutLength();
 
     @WamProperty(index = 4, type = WamType.STRING)
     Optional<String> aboutLocale();
