+++ modules/lib/src/main/java/com/github/auties00/cobalt/wam/event/AndroidMessageSendPerfEvent.java
@@ -20,7 +20,7 @@ import com.github.auties00.cobalt.wam.type.TypeOfGroupEnum;
 
 import java.time.Instant;
 import java.util.Optional;
-import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 @WhatsAppWebModule(moduleName = "WAWebAndroidMessageSendPerfWamEvent")
 @WamEvent(id = 1994, releaseWeight = 2000)
@@ -35,13 +35,13 @@ public interface AndroidMessageSendPerfEvent extends WamEventSpec {
     Optional<BotType> botType();
 
     @WamProperty(index = 32, type = WamType.INTEGER)
-    OptionalInt bucketedSenderKeyDistributionCountPercentage();
+    OptionalLong bucketedSenderKeyDistributionCountPercentage();
 
     @WamProperty(index = 35, type = WamType.INTEGER)
-    OptionalInt bucketedSenderKeyDistributionHashTime();
+    OptionalLong bucketedSenderKeyDistributionHashTime();
 
     @WamProperty(index = 36, type = WamType.INTEGER)
-    OptionalInt deviceCount();
+    OptionalLong deviceCount();
 
     @WamProperty(index = 26, type = WamType.ENUM)
     Optional<SizeBucket> deviceSizeBucket();
@@ -62,7 +62,7 @@ public interface AndroidMessageSendPerfEvent extends WamEventSpec {
     Optional<Boolean> fetchPrekeys();
 
     @WamProperty(index = 21, type = WamType.INTEGER)
-    OptionalInt fetchPrekeysPercentage();
+    OptionalLong fetchPrekeysPercentage();
 
     @WamProperty(index = 17, type = WamType.ENUM)
     Optional<ClientGroupSizeBucket> groupSizeBucket();
@@ -95,7 +95,7 @@ public interface AndroidMessageSendPerfEvent extends WamEventSpec {
     Optional<Boolean> isViewOnce();
 
     @WamProperty(index = 18, type = WamType.INTEGER)
-    OptionalInt jobsInQueue();
+    OptionalLong jobsInQueue();
 
     @WamProperty(index = 3, type = WamType.ENUM)
     Optional<MediaType> mediaType();
@@ -116,34 +116,34 @@ public interface AndroidMessageSendPerfEvent extends WamEventSpec {
     Optional<Boolean> networkWasDisconnected();
 
     @WamProperty(index = 37, type = WamType.INTEGER)
-    OptionalInt participantCount();
+    OptionalLong participantCount();
 
     @WamProperty(index = 34, type = WamType.INTEGER)
-    OptionalInt phoneCores();
+    OptionalLong phoneCores();
 
     @WamProperty(index = 28, type = WamType.BOOLEAN)
     Optional<Boolean> prekeysEligibleForPrallelProcessing();
 
     @WamProperty(index = 39, type = WamType.INTEGER)
-    OptionalInt receiverDeviceCount();
+    OptionalLong receiverDeviceCount();
 
     @WamProperty(index = 44, type = WamType.STRING)
     Optional<String> runningTasks();
 
     @WamProperty(index = 13, type = WamType.INTEGER)
-    OptionalInt sendCount();
+    OptionalLong sendCount();
 
     @WamProperty(index = 10, type = WamType.INTEGER)
-    OptionalInt sendRetryCount();
+    OptionalLong sendRetryCount();
 
     @WamProperty(index = 2, type = WamType.ENUM)
     Optional<ClientMessageSendStage> sendStage();
 
     @WamProperty(index = 40, type = WamType.INTEGER)
-    OptionalInt senderDeviceCount();
+    OptionalLong senderDeviceCount();
 
     @WamProperty(index = 23, type = WamType.INTEGER)
-    OptionalInt senderKeyDistributionCountPercentage();
+    OptionalLong senderKeyDistributionCountPercentage();
 
     @WamProperty(index = 25, type = WamType.BOOLEAN)
     Optional<Boolean> sessionsMissingWhenComposing();
@@ -152,7 +152,7 @@ public interface AndroidMessageSendPerfEvent extends WamEventSpec {
     Optional<ClientGroupSizeBucket> targetDeviceGroupSizeBucket();
 
     @WamProperty(index = 19, type = WamType.INTEGER)
-    OptionalInt threadsInExecution();
+    OptionalLong threadsInExecution();
 
     @WamProperty(index = 38, type = WamType.ENUM)
     Optional<TypeOfGroupEnum> typeOfGroup();
